Add hit_allies flag to fix is_checkmate
When generating the `targets` array in the king's move generator, until
now all fields where enemy pieces stand, were marked as not hittable.

A move that hits an enemy piece that is guarded by another was thus not
thrown out.

This adds a flag `hit_allies` that leads the generators to include moves
that guard pieces of the same color (technically the moves hit them).