Add undo_move and rename execute_move to do_move
This also fixes the bug that at the two places were we "manually" undid
a move a pawn promotion was not properly reset.
This should not have any influence on the correctness of the generator
as we are working on copies and as the figure type are not relevant at
those places. Though it could have still lead to some nasty bugs, if we
would have copy pasted those parts further.