Fix nasty bug that is_checkmate modified the board
This was because `cur_counter_move` was not undone when the mate was
declined and false returned directly. This was the case as earlier we
memcopied the board and did not need to worry about that.

Also terminate when using the negamax debug path and an empty list is
returned.