Change how `rate_move` works
`rate_move` should always rate the move as seen from the moving player.
I.e., a move which hits a piece should at least have the rating of
the hit piece as a rating (which is also the difference in the rating of
the board, before and after), and not return negative the difference
because the hitting player plays black.